[BEAM-12963]: reorganize files
