+++ playground/frontend/lib/playground_app.dart
@@ -17,12 +17,19 @@
  */
 
 import 'package:flutter/material.dart';
+import 'package:playground/config/theme.dart';
+import 'package:playground/pages/playground/playground_page.dart';
 
-class LightColor {
-  static const Color primaryBackground = Colors.white;
-  static const Color secondaryBackground = Color(0xFFFCFCFC);
-  static const Color grey = Color(0xFFE5E5E5);
-  static const Color grey1 = Color(0xFFA0A4AB);
-  static const Color text = Color(0xFF242639);
-  static const Color primary = Color(0xFFE74D1A);
+class PlaygroundApp extends StatelessWidget {
+  const PlaygroundApp({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Apache Beam Playground',
+      theme: theme,
+      home: const PlaygroundPage(),
+      debugShowCheckedModeBanner: false,
+    );
+  }
 }
