+++ playground/frontend/pubspec.yaml
@@ -48,11 +48,6 @@ dependencies:
     sdk: flutter
   provider: ^6.0.0
 
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-
 dev_dependencies:
   flutter_test:
     sdk: flutter
